Symlink /app/tmp to /tmp.
Many GOV.UK apps need this workaround because some third-party rubygems
assume that they can write to $APP_HOME/tmp.

We were previously doing this in individual apps' Dockerfiles, but it's
cleaner and less hassle to apply the workaround here.

(APP_HOME needs to be the path of the app's source directory, if I
understand correctly, so we can't just change that to point somewhere
writable.)