Add Ruby 3.1.3.
For now, we need to keep the previous patch version around in order to
switch over without friction. This is because the old (EC2)
build/deployment automation has to pin the exact Ruby version and that's
nontrivial to fix, so we just live with it for another couple of months
until we no longer have to care about staying compatible with it.

Once all the apps are on 3.1.3, we can delete the 3_1_2 file.

Checksums regenerated by `sha256sum *gz >SHA256SUMS`. They're now in
alphabetical order.