Update to Ruby 3.1.3 and 2.7.7.
For now, we need to keep the previous patch version around in order to
switch over without friction. This is because the old (EC2)
build/deployment automation has to pin the exact Ruby version and that's
nontrivial to fix, so we just live with it for another couple of months
until we no longer have to care about staying compatible with it.

Once all the apps are on 3.1.3 (and 2.7.7) - which ought to be within a
few days! - we can delete the 2_7_6 and 3_1_2 files.

Checksums regenerated by `sha256sum *gz >SHA256SUMS`. They're now in
alphabetical order.