Remove pkg-config package from initial build stage.
We don't really need it, and it turns out it breaks cross-compiling
unless we also pull in half a dozen other packages including Perl (for
weird Debiany reasons) so let's just drop it.

The build was working just fine without it. I only added it because I
thought it might be a tiny bit more future-proof / slightly less chance
of maintenance hassle in future, but that seems not to have panned out.