Add Ruby 3.2.0.
We need to add this obsolete patch version because the current (soon to
be obsolete) Puppet/EC2 system isn't going to be upgraded to Ruby 3.2.1
(checked with Platform Reliabilty team about this). So we're in a
somewhat strange situation where apps are being upgraded **to** 3.2.0 at
the moment even though 3.2.1 has been out for while.

(This is because govuk-puppet has become so time-consuming for PlatRel
to work with that 3.2.0 was superseded while they were working on
rolling it out, and they didn't fancy going through all that toil
again since we'll be moving to k8s within a month.)

Once Replatforming launches, we'll sort out the overly specific version
contraints/pinning so that we don't have to build old patch versions in
future. Fixing it before launch is hard because of govuk-puppet.