Add libyaml headers to builder image.
The `psych` gem needs to compile against libyaml.

Not sure why we're only running into this now (maybe Rails was happy
with our stock version of `psych` until recently?) but in any case it
seems sensible to include headers in the builder image for libraries
that we ship in our base image.