Add remaining missing headers to builder image.
We're not using any of these yet, but it makes sense to ship headers in
the builder image for all the libraries that we include in the base
image. Should reduce the chance of being caught out next time some
dependency starts requiring a native rubygem that depends on any of
those libs.