Fix shell script for local builds.
build.sh stopped working with the refactor of the GitHub Actions build.

Considered just dropping build.sh altogether, but it's useful to be able
to build locally (much faster feedback loop) and it's also peace of mind
to know that we could build/push without GitHub Actions if we had to.