Add libjsonnet++ for publishing-api Content Schemas.
Only publishing-api strictly needs this right now, but it's a small
library so adding it here is reasonable. Saves having to pull down the
Apt catalog in the pubapi build, and reduces the chance of
copypasta/cargo-culting extra steps from the pubapi Dockerfile.

See https://github.com/alphagov/publishing-api/pull/2788.