Keep a few more untagged "images" so we don't drop .atts.
We're not actually using the image signatures for anything yet, but
might as well avoid a potential subtle footshooter.

Most of the untagged "images" (not really images?) are the attestations
for image signing, so this just helps to ensure that if we start
checking signatures on container admission or something, we won't have
an unpleasant surprise if we ever needed to roll back to an older
base/builder image temporarily for whatever reason.