Set custom prompts using .bashrc and .irbrc
This is intended to remind developers which app and environment they're
targetting in interactive shells.

Technically the default shell (sh) in these container images is dash,
but I don't think anyone uses that interactively so I haven't tried to
customise it. zsh isn't installed, so we don't have to worry about that.

$APP_HOME is the homedir for the app user, so these should be in the right
place to be picked up by interactive shells / irb sessions.